requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
idna==3.3
Pillow==9.1.1
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
